pyproject.toml
[tool.poetry]
name = "crypto-bot"
version = "0.1.0"
description = ""
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = {extras = ["speedups"], version = "^3.10.10"}
msgspec = {extras = ["yaml"], version = "^0.18.6"}
python-dotenv = "^1.0.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.2"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 120
lint.select = ["ALL"]
lint.ignore = ["ANN401", "D"]
extend-exclude = ["local"]

[tool.mypy]
exclude = ["local"]
